Add functions for pegging a line in current file to a Git commit
Due to differences between the working tree and the Git HEAD commit, the
line number in the open file might correspond to different line number
in the HEAD commit (if it even exists).  These added functions implement
a best-effort mapping between the lines of the file in the working tree
and the lines in the same file in a commit (specified or HEAD).