Implement Git-pegging of bookmarks
Bookmarks can now be "pegged" against a Git commit: the bookmark stores
the Git commit hash and the closest line number representation within
the version of the file at that Git commit to the line of bookmark in
the working tree (which may be an approximation if the bookmark is
within a modified region of the file).

One known bug: the pegged-location system runs on on-disk file contents,
not the current contents of the buffer.  Unsaved changes in the buffer
can result in failure to locate the bookmark.