Fix Atom's broken drag-and-drop
When we are dragging, Atom has zapped our drag-and-drop data (`getData`
on the `dataTransfer` returns an empty string for our data type, even
though it had a valid value for the data type when we handled the
`dragstart` event on the `Draggable`) but kindly left the _data type_
in the data transfer type list.  We resort to using
`dragState.dragData` in this case to stop pulling out our hair.