Implement casefile recovery
Shared casefiles that have been deleted may now be searched by
typing at least 3 consecutive characters in the recovery search box,
then choosing a returned option and clicking "Recover".  The casefile
is recovered into the current casefile, which can then be shared again.
The "Share As" is automatically set to the path recovered.