Leverage `git blame` to find better bookmark anchors
Git's "blame" command, when run in porcelain mode, gives the best
possible information for the origin of a given line of code.  This
allows revision pegging to be approachable to more consumers because
A) `git blame` references the oldest commit with that same line, and
B) the older the commit, the more likely other users are to know it
(especially because some newer commits might not have been or ever be
published).  If "blame" can't give a valid commit, fall back on
previous behavior.