Protect more UI from locked models
Models reject actions while other actions are executing on the same
model (i.e. if the Promised state from an action is still pending).
This rejection comes with an error notification, so to avoid that,
disable the UI elements (mostly buttons) that can trigger the actions
when the associated model is locked (which condition the model
broadcasts to its bound views).