Fix bugs with persistent state
`catch` for a couple of `Promise`s returned from an action
implementation weren't returning state (putting the `Promise` in a
_fulfilled_ state with an actual _state_ value) or `throw`ing (putting
the `Promise` in a _rejected_ state), so the returned value
(`undefined`) was interpret as the fulfillment of the state promise,
which then became the state for the model.  The `catch` can always
just return an unmodified `state`.

Issues: #4