Attempt bookmark resolution with `git blame`
Because `git blame` is used to find the earliest version of the line
bookmarked (to make the reference as widely available as possible),
it should be used as the first-string resolution of the current line
location, falling back first to a diff with the content at the marked
commit, and finally to a search based on the original line number
marked.  The only change here is the introduction of the `git blame`
step.